refactor: 레이어드 아키텍처에 맞게 패키지 이동
